+++ lib/src/features/main/theme/bloc/theme_state.dart
@@ -3,8 +3,8 @@ part of 'theme_bloc.dart';
 class ThemeState {
   ThemeState._({required this.themeData});
 
-  factory ThemeState.initial(ThemeData themeData) {
+  factory ThemeState.initial(FluentThemeData themeData) {
     return ThemeState._(themeData: themeData);
   }
-  final ThemeData themeData;
+  final FluentThemeData themeData;
 }